GH-26 eclccPath setting has invalid cwd.
Fixes GH-26
