+++ src/clienttools/eclMeta.ts
@@ -6,7 +6,6 @@ import { scopedLogger } from "../util/logging";
 import { SAXStackParser, XMLNode } from "../util/saxParser";
 
 const logger = scopedLogger("clienttools/eclmeta");
-logger.info("Test");
 
 const _inspect = false;
 function inspect(obj: any, _id: string, known: any) {
